+++ functions/src/user/user.controller.ts
@@ -1,8 +1,10 @@
-import { Controller, Get, Post, Put, Delete, Param, Body, HttpCode, HttpStatus } from "@nestjs/common";
+import { Controller, Get, Post, Put, Delete, Param, Body, HttpCode, HttpStatus, UseGuards, ForbiddenException } from "@nestjs/common";
 import { ApiInternalServerErrorResponse, ApiTags } from "@nestjs/swagger";
 import { UserService } from "./user.service";
 import { CreateUser } from "./interfaces/create-user.interface";
 import { UpdateUser } from "./interfaces/update-user.interface";
+import { AuthJwtGuard } from "src/auth/guards/jwt-auth.guard";
+import { CurrentUser } from "src/auth/decorators/current-user.decorator";
 
 @Controller("user")
 @ApiTags("User")
@@ -30,13 +32,19 @@ export class UserController {
 
     @Put(":userId")
     @HttpCode(HttpStatus.OK)
-    async updateUser(@Param("userId") userId: string, @Body() updateUserDto: UpdateUser) {
+    @UseGuards(AuthJwtGuard)
+    async updateUser(@Param("userId") userId: string, @Body() updateUserDto: UpdateUser, @CurrentUser() user) {
+        if(userId !== user.userId)
+            throw new ForbiddenException("You can update only own user!");
         return await this.userService.updateUser(userId, updateUserDto);
     }
 
     @Delete(":userId")
     @HttpCode(HttpStatus.NO_CONTENT)
-    async deleteUser(@Param("userId") userId: string) {
+    @UseGuards(AuthJwtGuard)
+    async deleteUser(@Param("userId") userId: string, @CurrentUser() user) {
+        if(userId !== user.userId)
+            throw new ForbiddenException("You can delete only own user!");
         return await this.userService.deleteUser(userId);
     }
 }
